[MMCA-4977] Move the test class to the correct package, add tests and rename the val
